+++ example/knownfolder.dart
@@ -6,6 +6,28 @@ import 'dart:ffi';
 import 'package:ffi/ffi.dart';
 import 'package:win32/win32.dart';
 
+// Get the path of the temporary directory (typically %TEMP%)
+String getTemporaryPath() {
+  final buffer = allocate<Uint16>(count: MAX_PATH + 1).cast<Utf16>();
+  final length = GetTempPath(MAX_PATH, buffer);
+
+  if (length == 0) {
+    final error = GetLastError();
+    free(buffer);
+    throw WindowsException('$error');
+  } else {
+    var path = buffer.unpackString(MAX_PATH);
+
+    // GetTempPath adds a trailing backslash, but SHGetKnownFolderPath does not.
+    // Strip off trailing backslash for consistency with other methods here.
+    if (path[path.length - 1] == '\\') {
+      path = path.substring(0, path.length - 1);
+    }
+    free(buffer);
+    return path;
+  }
+}
+
 /// Get the path for a known Windows folder, using the classic (deprecated) API
 String getFolderPath() {
   var path = allocate<Uint16>(count: MAX_PATH);
@@ -21,29 +43,31 @@ String getFolderPath() {
 
 /// Get the path for a known Windows folder, using the modern API
 String getKnownFolderPath() {
-  final guidFolder = GUID.fromString(FOLDERID_Documents);
+  GUID knownFolderID = GUID.fromString(FOLDERID_Documents);
+  Pointer<IntPtr> pathPtrPtr = allocate<IntPtr>();
 
-  final ptr = allocate<IntPtr>();
-  final hr = SHGetKnownFolderPath(guidFolder.addressOf, 0, 0, ptr);
+  final hr = SHGetKnownFolderPath(
+      knownFolderID.addressOf, KF_FLAG_DEFAULT, NULL, pathPtrPtr);
 
-  if (SUCCEEDED(hr)) {
-    final path = Pointer<Utf16>.fromAddress(ptr.value).unpackString(MAX_PATH);
-
-    CoTaskMemFree(ptr.cast<Void>());
-
-    return path;
-  } else {
-    if (hr == E_FAIL) {
-      return ('E_FAIL');
-    } else if (hr == E_INVALIDARG) {
-      return ('E_INVALIDARG');
+  if (FAILED(hr)) {
+    if (hr == E_INVALIDARG || hr == E_FAIL) {
+      throw WindowsException('Invalid folder.');
     } else {
-      return ('error code 0x${hr.toUnsigned(32).toRadixString(16)}');
+      throw WindowsException('Unknown error.');
     }
   }
+
+  final pathPtr = Pointer<Utf16>.fromAddress(pathPtrPtr.value);
+  final path = pathPtr.unpackString(MAX_PATH);
+
+  CoTaskMemFree(pathPtr.cast());
+  free(pathPtrPtr);
+
+  return path;
 }
 
 void main() {
+  print('Temporary path is ${getTemporaryPath()}');
   print('SHGetFolderPath returned ${getFolderPath()}');
   print('SHGetKnownFolderPath returned ${getKnownFolderPath()}');
 }
