+++ tool/generate/winrt/IAsyncInfo.idl
@@ -1,0 +1,17 @@
+// vtable_start 6
+[uuid(00000036-0000-0000-C000-000000000046)]
+interface IAsyncInfo : IInspectable
+{
+    // Properties
+    [propget] HRESULT Id([out, retval] unsigned __int32 *id);
+
+    // provide a C++ overload for async status that doesn't rely on 
+    // the global definition of asyncstatus to support _HIDE_GLOBAL_ASYNC_STATUS
+    [propget] HRESULT Status([out, retval] AsyncStatus *status);
+
+    [propget] HRESULT ErrorCode([out,retval] HRESULT *errorCode);
+    
+    // Methods
+    HRESULT Cancel();
+    HRESULT Close();
+}
